fix: use int64 for block device sizes to fix 32-bit build
Disk sizes can exceed the max int32 on 32-bit targets (linux_386),
causing a compile-time overflow. Widen the Size fields and
formatBlockSize to int64 so the release builds for linux_386_sse2.

+++ pkg/actions/fs/blockdevice-diskutil_test.go
@@ -38,7 +38,7 @@ func TestBlockDevicesDiskutil(t *testing.T) {
 
 func TestFormatBlockSize(t *testing.T) {
 	tests := []struct {
-		bytes    int
+		bytes    int64
 		expected string
 	}{
 		{500, "500B"},
